Add Ui class
